% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_apitf.R
\name{read_apitf}
\alias{read_apitf}
\title{Read an APIT-F file into zero-copy R vectors}
\usage{
read_apitf(file, vars = NULL)
}
\arguments{
\item{file}{Path to \code{*.apitf}}

\item{vars}{Optional character vector of column indices or names.
\code{NULL} (default) maps every column.}
}
\value{
A named list of ALTREP vectors backed by the file on disk.
}
\description{
Read an APIT-F file into zero-copy R vectors
}
