% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tsv_with_schema.R
\name{read_tsv_with_schemata}
\alias{read_tsv_with_schemata}
\alias{test_rtsv}
\alias{test_rtsv_ynq}
\title{Read a TSV with schemata}
\usage{
read_tsv_with_schemata(file.tsv, Schemata, nThread = 2L)

test_rtsv(nThread = 1L)

test_rtsv_ynq(nThread = 1L)
}
\arguments{
\item{file.tsv}{A tsv file.}

\item{Schemata}{A schemata produced by \code{\link{field_schemata}}.}

\item{nThread}{Number of OpenMP threads to use.}
}
\value{
The number of rows.
}
\description{
Read a TSV with schemata
}
