% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aptif.R
\name{write_apitf}
\alias{write_apitf}
\title{Convert delimited text to APIT-F}
\usage{
write_apitf(schema, data, dest, delim = "\\t", tax_year = 2024L, nThread = 1L)
}
\arguments{
\item{schema, data, dest}{Paths (character scalar)}

\item{delim}{Single-character delimiter (default \code{"\t"})}

\item{tax_year}{4-digit integer (default 2024)}

\item{nThread}{Number of threads to use}
}
\value{
Invisibly returns \code{dest}
}
\description{
Convert delimited text to APIT-F
}
