% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_scehemata.R
\name{field_schemata}
\alias{field_schemata}
\alias{field_schema}
\title{Create a field schemata list}
\usage{
field_schemata(..., FILE.YAML = NULL)

field_schema(
  type = c("bit1", "bit2", "bit16", "int100", "int32", "int64", "double", "string"),
  required = FALSE,
  i_min = NULL,
  i_max = NULL,
  valid = NULL
)
}
\arguments{
\item{...}{A named list of calls to `field_schema`.}

\item{FILE.YAML}{A filename, the YAML file which specifies the schema.}

\item{type}{One of the specified types.}

\item{required}{(bool) whether the name is required (i.e. whether its absence should raise an error).}

\item{i_min, i_max}{If the type is an integer type, the range of allowable integers.}

\item{valid}{A vector of strings that are valid for this field.}
}
\value{
A list, indicating the schema for each column.
}
\description{
Create a field schemata list
}
