% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tsv_with_schema.R
\name{read_tsv_with_schemata}
\alias{read_tsv_with_schemata}
\alias{test_rtsv}
\title{Read a TSV with schemata}
\usage{
read_tsv_with_schemata(file.tsv, Schemata, nThread = 2L)

test_rtsv()
}
\arguments{
\item{file.tsv}{A tsv file.}

\item{Schemata}{A schemata produced by \code{\link{field_schemata}}.}
}
\value{
The number of rows.
}
\description{
Read a TSV with schemata
}
